+++ asa_application/statistics.h
@@ -3,8 +3,8 @@
 #include <Qt>
 #include <QList>
 
-void add_value_to_stats(uint param, int new_value);
-QList<int> get_list_from_param(uint param);
+void add_value_to_stats(uint param, float new_value);
+QList<float> get_list_from_param(uint param);
 QString get_last_value_from_param(uint param);
